fix по ошибке с кодом ответа

+++ main-service/src/main/java/ru/practicum/dto/event/EventDTO.java
@@ -210,7 +210,7 @@ public enum EventDTO {
             LocalDateTime eventDate;
             LocationDto location;
             Boolean paid;
-            Integer participantLimit;
+            Long participantLimit;
             Boolean requestModeration;
             StateActionUser stateAction;
             String title;
@@ -233,7 +233,7 @@ public enum EventDTO {
             LocalDateTime eventDate;
             LocationDto location;
             Boolean paid;
-            Integer participantLimit;
+            Long participantLimit;
             Boolean requestModeration;
             StateActionAdmin stateAction;
             String title;
